// reportsModule/GalvanizingСranesheetReport/types.ts
import { Nullable } from '@types';
import { $Enums, LineItemUnitNote } from '@prisma/client';
import ProductionLocation = $Enums.ProductionLocation;

export type CraneTableColumns = {
  id: string;
  orderNumber: string;
  dateCreated: Nullable<string>;
  PO: string;
  customer: string;
  tracking: string;
  lineItemType: ProductionLocation | string;
  lineItemDescription: string;
  pieces: number;
  status: string;
  rackNumber: string;
  rackType: string;
  firstDipKettleInTime: Nullable<string>;
  firstDipKettleOutTime: Nullable<string>;
  firstDipTimeRange: Nullable<string>;
  secondDipKettleInTime: Nullable<string>;
  secondDipKettleOutTime: Nullable<string>;
  secondDipTimeRange: Nullable<string>;
  rackWeight: number;
  firstDipWeightIn: number;
  firstDipWeightOut: number;
  secondDipWeightIn: number;
  secondDipWeightOut: number;
  zincPickUp: number;
  notes: string;
};

export type CraneReport = {
  id: string;
  orderNumber: string;
  weight: number;
  rackCreatedAt: string;
  customerName: string;
  customerPO: string;
  tracking: string;
  lineItemType: ProductionLocation | string;
  description: string;
  quantity: number;
  status: string;
  rackType: string;
  rackNumber: string;
  firstDipKettleInTime: Nullable<string>;
  firstDipKettleOutTime: Nullable<string>;
  firstDipTimeRange: Nullable<string>;
  secondDipKettleInTime: Nullable<string>;
  secondDipKettleOutTime: Nullable<string>;
  secondDipTimeRange: Nullable<string>;
  rackWeight: number;
  firstDipWeightIn: number;
  firstDipWeightOut: number;
  secondDipWeightIn: number;
  secondDipWeightOut: number;
  zincPickUp: number;
  notes: LineItemUnitNote[];
};